requirements.txt
flask==3.0.0
python-telegram-bot==20.7
openai==1.14.0
httpx==0.27.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-speech==1.37.0
msrest==0.7.1
gunicorn==21.2.0
